Move the NoHandler to the grid web package
